+++ web/js/MapUI/Components/UpdateProjection/UpdateProjection.js
@@ -1,48 +1,312 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import OlLayerGroup from 'ol/layer/Group';
+import {
+  each as lodashEach,
+  get as lodashGet,
+} from 'lodash';
+import {
+  fitToLeadingExtent,
+  updateMapUI,
+} from '../../../modules/map/actions';
+import { getLeadingExtent } from '../../../modules/map/util';
+import {
+  getLayers,
+  getActiveLayers,
+} from '../../../modules/layers/selectors';
+import { getSelectedDate } from '../../../modules/date/selectors';
+import util from '../../../util/util';
+import { fly } from '../../../map/util';
+import * as layerConstants from '../../../modules/layers/constants';
+import * as compareConstants from '../../../modules/compare/constants';
+import * as paletteConstants from '../../../modules/palettes/constants';
+import * as vectorStyleConstants from '../../../modules/vector-styles/constants';
+import { LOCATION_POP_ACTION } from '../../../redux-location-state-customs';
+import { EXIT_ANIMATION, STOP_ANIMATION } from '../../../modules/animation/constants';
+import { SET_SCREEN_INFO } from '../../../modules/screen-size/constants';
 
 const UpdateProjection = (props) => {
   const {
+    action,
+    activeLayers,
+    compare,
+    compareMapUi,
+    compareMode,
     config,
+    dateCompareState,
+    fitToLeadingExtent,
+    getGranuleOptions,
+    isMobile,
+    layerState,
     map,
-  } = props
+    models,
+    preloadForCompareMode,
+    proj,
+    projectionTrigger,
+    updateExtent,
+    updateLayerVisibilities,
+    updateMapUI,
+    ui,
+  } = props;
 
   useEffect(() => {
-    console.log('update projection');
-    console.log(map)
-  })
+    actionSwitch();
+  }, [action]);
 
-  function updateProjection(start) {
-    if (uiCopy.selected) {
-    // Keep track of center point on projection switch
-      uiCopy.selected.previousCenter = uiCopy.selected.center;
-      hideMap(uiCopy.selected);
+  useEffect(() => {
+    if (projectionTrigger === 1) {
+      updateProjection(true);
+    } else if (projectionTrigger > 1) {
+      updateProjection();
     }
-    uiCopy.selected = uiCopy.proj[proj.id];
-    const map = uiCopy.selected;
+  }, [projectionTrigger]);
+
+  const actionSwitch = () => {
+    switch (action.type) {
+      case STOP_ANIMATION:
+      case EXIT_ANIMATION:
+        return onStopAnimation();
+      case LOCATION_POP_ACTION: {
+        const newState = util.fromQueryString(action.payload.search);
+        const extent = lodashGet(map, 'extent');
+        const rotate = lodashGet(map, 'rotation') || 0;
+        setTimeout(() => {
+          updateProjection();
+          if (newState.v && !newState.e && extent) {
+            flyToNewExtent(extent, rotate);
+          }
+        }, 200); return;
+      }
+      case layerConstants.UPDATE_GRANULE_LAYER_OPTIONS: {
+        const granuleOptions = {
+          id: action.id,
+          reset: null,
+        };
+        return reloadLayers(granuleOptions);
+      }
+      case layerConstants.RESET_GRANULE_LAYER_OPTIONS: {
+        const granuleOptions = {
+          id: action.id,
+          reset: action.id,
+        };
+        return reloadLayers(granuleOptions);
+      }
+      case compareConstants.CHANGE_STATE:
+        if (compareMode === 'spy') {
+          return reloadLayers();
+        }
+        return;
+      case layerConstants.REORDER_LAYERS:
+      case layerConstants.REORDER_OVERLAY_GROUPS:
+      case compareConstants.TOGGLE_ON_OFF:
+      case compareConstants.CHANGE_MODE:
+        reloadLayers();
+        preloadForCompareMode();
+        return;
+      case layerConstants.TOGGLE_OVERLAY_GROUPS:
+        return reloadLayers();
+      case paletteConstants.SET_THRESHOLD_RANGE_AND_SQUASH:
+      case paletteConstants.SET_CUSTOM:
+      case paletteConstants.SET_DISABLED_CLASSIFICATION:
+      case paletteConstants.CLEAR_CUSTOM:
+      case layerConstants.ADD_LAYERS_FOR_EVENT:
+        return setTimeout(reloadLayers, 100);
+      case vectorStyleConstants.SET_FILTER_RANGE:
+      case vectorStyleConstants.SET_VECTORSTYLE:
+      case vectorStyleConstants.CLEAR_VECTORSTYLE:
+      case SET_SCREEN_INFO:
+        return onResize();
+      default:
+        break;
+    }
+  };
+
+  const flyToNewExtent = function(extent, rotation) {
+    const coordinateX = extent[0] + (extent[2] - extent[0]) / 2;
+    const coordinateY = extent[1] + (extent[3] - extent[1]) / 2;
+    const coordinates = [coordinateX, coordinateY];
+    const resolution = ui.selected.getView().getResolutionForExtent(extent);
+    const zoom = ui.selected.getView().getZoomForResolution(resolution);
+    // Animate to extent, zoom & rotate:
+    // Don't animate when an event is selected (Event selection already animates)
+    return fly(ui.selected, proj, coordinates, zoom, rotation);
+  };
+
+  /**
+ * Create a Layergroup given the date and layerGroups
+ */
+  async function getCompareLayerGroup([compareActiveString, compareDateString], state, granuleOptions) {
+    const { createLayer } = ui;
+    const compareSideLayers = getActiveLayers(state, compareActiveString);
+    const layers = getLayers(state, { reverse: true }, compareSideLayers)
+      .map(async (def) => {
+        const options = {
+          ...getGranuleOptions(state, def, compareActiveString, granuleOptions),
+          date: getSelectedDate(dateCompareState, compareDateString),
+          group: compareActiveString,
+        };
+        return createLayer(def, options);
+      });
+    const compareLayerGroup = await Promise.all(layers);
+
+    return new OlLayerGroup({
+      layers: compareLayerGroup,
+      date: getSelectedDate(dateCompareState, compareDateString),
+      group: compareActiveString,
+    });
+  }
+
+  /**
+  * Remove Layers from map
+  *
+  * @method clearLayers
+  * @static
+  *
+  * @param {object} map - Openlayers Map obj
+  *
+  * @returns {void}
+  */
+  const clearLayers = function() {
+    const activeLayers = ui.selected
+      .getLayers()
+      .getArray()
+      .slice(0);
+    lodashEach(activeLayers, (mapLayer) => {
+      ui.selected.removeLayer(mapLayer);
+    });
+    ui.cache.clear();
+  };
+
+  /**
+   * @method reloadLayers
+   *
+   * @param {object} map - Openlayers Map obj
+   * @param {Object} granuleOptions (optional: only used for granule layers)
+   *   @param {Boolean} granuleDates - array of granule dates
+   *   @param {Boolean} id - layer id
+   * @returns {void}
+   */
+  async function reloadLayers(granuleOptions) {
+    console.log('reloading layers');
+    const map = ui.selected;
+    const { createLayer } = ui;
+
+    if (!config.features.compare || !compare.active) {
+      const compareMapDestroyed = !compare.active && compareMapUi.active;
+      if (compareMapDestroyed) {
+        compareMapUi.destroy();
+      }
+      clearLayers();
+      const defs = getLayers(layerState, { reverse: true });
+      const layerPromises = defs.map((def) => {
+        const options = getGranuleOptions(layerState, def, compare.activeString, granuleOptions);
+        return createLayer(def, options);
+      });
+      const createdLayers = await Promise.all(layerPromises);
+      lodashEach(createdLayers, (l) => { map.addLayer(l); });
+    } else {
+      const stateArray = [['active', 'selected'], ['activeB', 'selectedB']];
+      if (compare && !compare.isCompareA && compare.mode === 'spy') {
+        stateArray.reverse(); // Set Layer order based on active A|B group
+      }
+      clearLayers();
+      const stateArrayGroups = stateArray.map(async (arr) => getCompareLayerGroup(arr, layerState, granuleOptions));
+      const compareLayerGroups = await Promise.all(stateArrayGroups);
+      compareLayerGroups.forEach((layerGroup) => map.addLayer(layerGroup));
+      compareMapUi.create(map, compare.mode);
+    }
+    updateLayerVisibilities();
+  }
+
+  const onStopAnimation = function() {
+    const needsRefresh = activeLayers.some(({ type }) => type === 'granule' || type === 'vector');
+    if (needsRefresh) {
+      // The SELECT_DATE and STOP_ANIMATION actions happen back to back and both
+      // try to modify map layers asynchronously so we need to set a timeout to allow
+      // the updateDate() function to complete before trying to call reloadLayers() here
+      setTimeout(reloadLayers, 100);
+    }
+  };
+
+  /**
+ * When page is resized set for mobile or desktop
+ *
+ * @method onResize
+ * @static
+ *
+ * @returns {void}
+ */
+  function onResize() {
+    const map = ui.selected;
+    if (isMobile) {
+      map.removeControl(map.wv.scaleImperial);
+      map.removeControl(map.wv.scaleMetric);
+    } else {
+      map.addControl(map.wv.scaleImperial);
+      map.addControl(map.wv.scaleMetric);
+    }
+  }
+
+  /**
+ * Show Map
+ *
+ * @method showMap
+ * @static
+ *
+ * @param {object} map - Openlayers Map obj
+ *
+ * @returns {void}
+ */
+  function showMap(map) {
+    document.getElementById(`${map.getTarget()}`).style.display = 'block';
+  }
+
+  /**
+ * Hide Map
+ *
+ * @method hideMap
+ * @static
+ *
+ * @param {object} map - Openlayers Map obj
+ *
+ * @returns {void}
+ */
+  function hideMap(map) {
+    document.getElementById(`${map.getTarget()}`).style.display = 'none';
+  }
+
+
+  const updateProjection = (start) => {
+    console.log('3. Updating Projection');
+    if (ui.selected) {
+      // Keep track of center point on projection switch
+      ui.selected.previousCenter = ui.selected.center;
+      hideMap(ui.selected);
+    }
+    ui.selected = ui.proj[proj.id];
+    const map = ui.selected;
 
     const isProjectionRotatable = proj.id !== 'geographic' && proj.id !== 'webmerc';
     const currentRotation = isProjectionRotatable ? map.getView().getRotation() : 0;
     const rotationStart = isProjectionRotatable ? models.map.rotation : 0;
+    const rotation = start ? rotationStart : currentRotation;
 
-    store.dispatch({
-      type: UPDATE_MAP_UI,
-      ui: uiCopy,
-      rotation: start ? rotationStart : currentRotation,
-    });
+    updateMapUI(ui, rotation);
+
     reloadLayers();
 
     // If the browser was resized, the inactive map was not notified of
     // the event. Force the update no matter what and reposition the center
     // using the previous value.
     showMap(map);
+
     map.updateSize();
 
-    if (uiCopy.selected.previousCenter) {
-      uiCopy.selected.setCenter(uiCopy.selected.previousCenter);
+    if (ui.selected.previousCenter) {
+      ui.selected.setCenter(ui.selected.previousCenter);
     }
-    // This is awkward and needs a refactoring
+
     if (start) {
       const projId = proj.selected.id;
       let extent = null;
@@ -54,7 +318,7 @@ const UpdateProjection = (props) => {
         callback = () => {
           const view = map.getView();
           const extent = view.calculateExtent(map.getSize());
-          store.dispatch({ type: FITTED_TO_LEADING_EXTENT, extent });
+          fitToLeadingExtent(extent);
         };
       }
       if (projId !== 'geographic') {
@@ -74,22 +338,64 @@ const UpdateProjection = (props) => {
     }
     updateExtent();
     onResize();
-    // handleActiveMapMarker(start);
-  }
+  };
 
-  return (
-    <>
-    </>
-  );
-}
+  return null;
+};
 
 const mapStateToProps = (state) => {
-  const { proj, map } = state
+  const {
+    proj, map, screenSize, layers, compare, date,
+  } = state;
+  const layerState = { layers, compare, proj };
+  const isMobile = screenSize.isMobileDevice;
+  const dateCompareState = { date, compare };
+  const activeLayers = getActiveLayers(state);
+  const compareMode = compare.mode;
   return {
-    proj, map
-  }
-}
+    activeLayers,
+    compare,
+    compareMode,
+    dateCompareState,
+    isMobile,
+    layerState,
+    proj,
+    map,
+  };
+};
+
+const mapDispatchToProps = (dispatch) => ({
+  fitToLeadingExtent: (extent) => {
+    dispatch(fitToLeadingExtent(extent));
+  },
+  updateMapUI: (ui, rotation) => {
+    dispatch(updateMapUI(ui, rotation));
+  },
+});
 
 export default connect(
   mapStateToProps,
-)(UpdateProjection)
+  mapDispatchToProps,
+)(UpdateProjection);
+
+UpdateProjection.propTypes = {
+  action: PropTypes.object,
+  activeLayers: PropTypes.array,
+  compare: PropTypes.object,
+  compareMapUi: PropTypes.object,
+  config: PropTypes.object,
+  dateCompareState: PropTypes.object,
+  fitToLeadingExtent: PropTypes.func,
+  getGranuleOptions: PropTypes.func,
+  isMobile: PropTypes.bool,
+  layerState: PropTypes.object,
+  map: PropTypes.object,
+  models: PropTypes.object,
+  preloadForCompareMode: PropTypes.func,
+  proj: PropTypes.object,
+  projectionTrigger: PropTypes.number,
+  ui: PropTypes.object,
+  updateExtent: PropTypes.func,
+  updateLayerVisibilities: PropTypes.func,
+  updateMapUi: PropTypes.func,
+};
